Oops... commit tests as well
